requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
SpeechRecognition==3.10.0
pyaudio==0.2.14
pyttsx3==2.90
Pillow==10.0.1
# Tkinter通常随Python自带，但Pillow用于图片处理

